+++ Src/Sla/DECCORE/RuleSignShift.cs
@@ -68,7 +68,7 @@ namespace Sla.DECCORE
             data.opSetOpcode(shiftOp, OpCode.CPUI_INT_SRIGHT);
             Varnode uniqueVn = data.newUniqueOut(inVn.getSize(), shiftOp);
             data.opSetInput(op, uniqueVn, 0);
-            data.opSetInput(op, data.newConstant(inVn.getSize(), Globals.calc_mask(inVn.getSize())), 1);
+            data.opSetInput(op, data.newConstant(inVn.getSize(), Globals.calc_mask((uint)inVn.getSize())), 1);
             data.opSetOpcode(op, OpCode.CPUI_INT_MULT);
             data.opSetInput(shiftOp, inVn, 0);
             data.opSetInput(shiftOp, constVn, 1);
